refactor: extract thumbnails into js/thumbnails.js
